+++ internal/cmd/floatingip/labels.go
@@ -15,7 +15,7 @@ var labelCmds = base.LabelCmds{
 	ShortDescriptionRemove: "Remove a label from an Floating IP",
 	NameSuggestions:        func(c hcapi2.Client) func() []string { return c.FloatingIP().Names },
 	LabelKeySuggestions:    func(c hcapi2.Client) func(idOrName string) []string { return c.FloatingIP().LabelKeys },
-	FetchLabels: func(ctx context.Context, client hcapi2.Client, idOrName string) (map[string]string, int, error) {
+	FetchLabels: func(ctx context.Context, client hcapi2.Client, idOrName string) (map[string]string, int64, error) {
 		floatingIP, _, err := client.FloatingIP().Get(ctx, idOrName)
 		if err != nil {
 			return nil, 0, err
@@ -25,7 +25,7 @@ var labelCmds = base.LabelCmds{
 		}
 		return floatingIP.Labels, floatingIP.ID, nil
 	},
-	SetLabels: func(ctx context.Context, client hcapi2.Client, id int, labels map[string]string) error {
+	SetLabels: func(ctx context.Context, client hcapi2.Client, id int64, labels map[string]string) error {
 		opts := hcloud.FloatingIPUpdateOpts{
 			Labels: labels,
 		}
