+++ internal/cmd/placementgroup/labels.go
@@ -15,7 +15,7 @@ var LabelCmds = base.LabelCmds{
 	ShortDescriptionRemove: "Remove a label from a placement group",
 	NameSuggestions:        func(c hcapi2.Client) func() []string { return c.PlacementGroup().Names },
 	LabelKeySuggestions:    func(c hcapi2.Client) func(idOrName string) []string { return c.PlacementGroup().LabelKeys },
-	FetchLabels: func(ctx context.Context, client hcapi2.Client, idOrName string) (map[string]string, int, error) {
+	FetchLabels: func(ctx context.Context, client hcapi2.Client, idOrName string) (map[string]string, int64, error) {
 		placementGroup, _, err := client.PlacementGroup().Get(ctx, idOrName)
 		if err != nil {
 			return nil, 0, err
@@ -25,7 +25,7 @@ var LabelCmds = base.LabelCmds{
 		}
 		return placementGroup.Labels, placementGroup.ID, nil
 	},
-	SetLabels: func(ctx context.Context, client hcapi2.Client, id int, labels map[string]string) error {
+	SetLabels: func(ctx context.Context, client hcapi2.Client, id int64, labels map[string]string) error {
 		opts := hcloud.PlacementGroupUpdateOpts{
 			Labels: labels,
 		}
