+++ internal/cmd/placementgroup/labels_test.go
@@ -27,7 +27,7 @@ func TestAddLabel(t *testing.T) {
 		Name:    "my Placement Group",
 		Created: time.Now(),
 		Labels:  map[string]string{"key": "value"},
-		Servers: []int{4711, 4712},
+		Servers: []int64{4711, 4712},
 		Type:    hcloud.PlacementGroupTypeSpread,
 	}
 
@@ -62,7 +62,7 @@ func TestRemoveLabel(t *testing.T) {
 		Name:    "my Placement Group",
 		Created: time.Now(),
 		Labels:  map[string]string{"key": "value"},
-		Servers: []int{4711, 4712},
+		Servers: []int64{4711, 4712},
 		Type:    hcloud.PlacementGroupTypeSpread,
 	}
