+++ internal/cmd/primaryip/labels.go
@@ -15,7 +15,7 @@ var labelCmds = base.LabelCmds{
 	ShortDescriptionRemove: "Remove a label from a Primary IP",
 	NameSuggestions:        func(c hcapi2.Client) func() []string { return c.PrimaryIP().Names },
 	LabelKeySuggestions:    func(c hcapi2.Client) func(idOrName string) []string { return c.PrimaryIP().LabelKeys },
-	FetchLabels: func(ctx context.Context, client hcapi2.Client, idOrName string) (map[string]string, int, error) {
+	FetchLabels: func(ctx context.Context, client hcapi2.Client, idOrName string) (map[string]string, int64, error) {
 		primaryIP, _, err := client.PrimaryIP().Get(ctx, idOrName)
 		if err != nil {
 			return nil, 0, err
@@ -25,7 +25,7 @@ var labelCmds = base.LabelCmds{
 		}
 		return primaryIP.Labels, primaryIP.ID, nil
 	},
-	SetLabels: func(ctx context.Context, client hcapi2.Client, id int, labels map[string]string) error {
+	SetLabels: func(ctx context.Context, client hcapi2.Client, id int64, labels map[string]string) error {
 		opts := hcloud.PrimaryIPUpdateOpts{
 			Labels: &labels,
 		}
