+++ internal/hcapi2/mock/zz_iso_client_mock.go
@@ -82,7 +82,7 @@ func (mr *MockISOClientMockRecorder) Get(arg0, arg1 interface{}) *gomock.Call {
 }
 
 // GetByID mocks base method.
-func (m *MockISOClient) GetByID(arg0 context.Context, arg1 int) (*hcloud.ISO, *hcloud.Response, error) {
+func (m *MockISOClient) GetByID(arg0 context.Context, arg1 int64) (*hcloud.ISO, *hcloud.Response, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetByID", arg0, arg1)
 	ret0, _ := ret[0].(*hcloud.ISO)
