+++ internal/hcapi2/zz_firewall_client_base.go
@@ -4,7 +4,6 @@ package hcapi2
 
 import (
 	"context"
-
 	"github.com/hetznercloud/hcloud-go/v2/hcloud"
 )
 
@@ -16,7 +15,7 @@ type FirewallClientBase interface {
 	Create(context.Context, hcloud.FirewallCreateOpts) (hcloud.FirewallCreateResult, *hcloud.Response, error)
 	Delete(context.Context, *hcloud.Firewall) (*hcloud.Response, error)
 	Get(context.Context, string) (*hcloud.Firewall, *hcloud.Response, error)
-	GetByID(context.Context, int) (*hcloud.Firewall, *hcloud.Response, error)
+	GetByID(context.Context, int64) (*hcloud.Firewall, *hcloud.Response, error)
 	GetByName(context.Context, string) (*hcloud.Firewall, *hcloud.Response, error)
 	List(context.Context, hcloud.FirewallListOpts) ([]*hcloud.Firewall, *hcloud.Response, error)
 	RemoveResources(context.Context, *hcloud.Firewall, []hcloud.FirewallResource) ([]*hcloud.Action, *hcloud.Response, error)
