+++ internal/hcapi2/zz_image_client_base.go
@@ -4,7 +4,6 @@ package hcapi2
 
 import (
 	"context"
-
 	"github.com/hetznercloud/hcloud-go/v2/hcloud"
 )
 
@@ -15,7 +14,7 @@ type ImageClientBase interface {
 	ChangeProtection(context.Context, *hcloud.Image, hcloud.ImageChangeProtectionOpts) (*hcloud.Action, *hcloud.Response, error)
 	Delete(context.Context, *hcloud.Image) (*hcloud.Response, error)
 	Get(context.Context, string) (*hcloud.Image, *hcloud.Response, error)
-	GetByID(context.Context, int) (*hcloud.Image, *hcloud.Response, error)
+	GetByID(context.Context, int64) (*hcloud.Image, *hcloud.Response, error)
 	GetByName(context.Context, string) (*hcloud.Image, *hcloud.Response, error)
 	GetByNameAndArchitecture(context.Context, string, hcloud.Architecture) (*hcloud.Image, *hcloud.Response, error)
 	GetForArchitecture(context.Context, string, hcloud.Architecture) (*hcloud.Image, *hcloud.Response, error)
