+++ internal/hcapi2/zz_iso_client_base.go
@@ -4,7 +4,6 @@ package hcapi2
 
 import (
 	"context"
-
 	"github.com/hetznercloud/hcloud-go/v2/hcloud"
 )
 
@@ -13,7 +12,7 @@ type ISOClientBase interface {
 	All(context.Context) ([]*hcloud.ISO, error)
 	AllWithOpts(context.Context, hcloud.ISOListOpts) ([]*hcloud.ISO, error)
 	Get(context.Context, string) (*hcloud.ISO, *hcloud.Response, error)
-	GetByID(context.Context, int) (*hcloud.ISO, *hcloud.Response, error)
+	GetByID(context.Context, int64) (*hcloud.ISO, *hcloud.Response, error)
 	GetByName(context.Context, string) (*hcloud.ISO, *hcloud.Response, error)
 	List(context.Context, hcloud.ISOListOpts) ([]*hcloud.ISO, *hcloud.Response, error)
 }
