+++ internal/hcapi2/zz_ssh_key_client_base.go
@@ -4,7 +4,6 @@ package hcapi2
 
 import (
 	"context"
-
 	"github.com/hetznercloud/hcloud-go/v2/hcloud"
 )
 
@@ -16,7 +15,7 @@ type SSHKeyClientBase interface {
 	Delete(context.Context, *hcloud.SSHKey) (*hcloud.Response, error)
 	Get(context.Context, string) (*hcloud.SSHKey, *hcloud.Response, error)
 	GetByFingerprint(context.Context, string) (*hcloud.SSHKey, *hcloud.Response, error)
-	GetByID(context.Context, int) (*hcloud.SSHKey, *hcloud.Response, error)
+	GetByID(context.Context, int64) (*hcloud.SSHKey, *hcloud.Response, error)
 	GetByName(context.Context, string) (*hcloud.SSHKey, *hcloud.Response, error)
 	List(context.Context, hcloud.SSHKeyListOpts) ([]*hcloud.SSHKey, *hcloud.Response, error)
 	Update(context.Context, *hcloud.SSHKey, hcloud.SSHKeyUpdateOpts) (*hcloud.SSHKey, *hcloud.Response, error)
